uuid: make integer arithmetic conversion explicit
Upcast multiplication to 64-bit integer arithmetic.

+++ cpukit/libmisc/uuid/gen_uuid.c
@@ -402,7 +402,7 @@ try_again:
 		last = tv;
 	}
 
-	clock_reg = tv.tv_usec*10 + adjustment;
+	clock_reg = ((uint64_t) tv.tv_usec)*10 + adjustment;
 	clock_reg += ((uint64_t) tv.tv_sec)*10000000;
 	clock_reg += (((uint64_t) 0x01B21DD2) << 32) + 0x13814000;
